frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["uday kiran <udaykiranchintu16@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"
psycopg2 = "^2.9.9"
google-cloud-storage = "^2.18.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
streamlit-shadcn-ui = "^0.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
